Fix some issues with slopes, and some refactoring
 - Fix Hitboxes sometimes colliding at the bottom of a ceiling slope

 - Fix ceiling slopes not preserving downward momentum

 - Remove references to AWT (not supported by LWJGL3 on MacOS)

 - Improve Logic construction (Level should not be required at construction)

+++ src/engine/game/tiles/Tile.java
@@ -84,7 +84,7 @@ public abstract class Tile {
      * @param x Position in world units.
      * @return Tile edge in world units.
      */
-    public static float getLeft(float x) {
+    public static int getLeft(float x) {
         return (int) (x / WIDTH);
     }
 
@@ -107,7 +107,7 @@ public abstract class Tile {
      * @param y Position in world units.
      * @return Tile edge in world units.
      */
-    public static float getTop(float y) {
+    public static int getTop(float y) {
         return (int) (y / HEIGHT);
     }
